change error recovery priority
Skipping tokens is generally better than inserting `}` because the latter tends to close all open scopes and then skip the rest of the tokens. However, we have to make sure not to insert an infinite sequence of `_;` at block-level. To achieve that, we never add `_` after `;`, which should be safe.